Split ReportingControllerSpec in two files
